% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{add_clustree_colour_scale}
\alias{add_clustree_colour_scale}
\title{Add a clustree colour/fill scale}
\usage{
add_clustree_colour_scale(gg, values, aesthetic = c("colour", "fill"))
}
\arguments{
\item{gg}{ggplot object to add the scale to}

\item{values}{vector of values being mapped to the aesthetic}

\item{aesthetic}{either "colour" or "fill"}
}
\description{
Adds a discrete scale using the clustree palette or, for numeric data
(e.g. gene expression), a continuous viridis scale, matching the
aesthetic used for that data.
}
\keyword{internal}
