% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{clustree_edge_colour}
\alias{clustree_edge_colour}
\title{Default flat edge colour for clustree plots}
\format{
An object of class \code{character} of length 1.
}
\usage{
clustree_edge_colour
}
\description{
Default flat edge colour for clustree plots
}
\keyword{internal}
