% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{clustree_palette}
\alias{clustree_palette}
\title{clustree colour palette}
\format{
An object of class \code{character} of length 20.
}
\usage{
clustree_palette
}
\description{
A categorical colour palette (matplotlib's "tab20", as used by packages
such as \code{pyclustree}) used as the default discrete node colour scale for
clustering tree plots. A qualitative palette of muted, evenly spaced hues
is easier to visually separate than \code{ggplot2}'s default (highly
saturated) hue palette.
}
\keyword{internal}
