% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{clustree_palette_n}
\alias{clustree_palette_n}
\title{Get \code{n} colours from the clustree palette}
\usage{
clustree_palette_n(n)
}
\arguments{
\item{n}{number of colours required}
}
\description{
Recycles \link{clustree_palette} if there are more levels than colours.
}
\keyword{internal}
