% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{sort_clusters}
\alias{sort_clusters}
\title{Sort cluster labels}
\usage{
sort_clusters(clusters)
}
\arguments{
\item{clusters}{vector of cluster labels}
}
\value{
sorted vector of cluster labels
}
\description{
Sort a vector of cluster labels numerically when every label can be
interpreted as a number and lexicographically otherwise.
}
\keyword{internal}
